feat(modules.profile): add cameraSettings
