+++ packages/modules.profile/src/ui/Header.tsx
@@ -9,6 +9,7 @@ const menuLabels = [
   'Безопасность',
   'Уведомления',
   'Эффекты',
+  'Настройки видео',
   'Отчёт',
 ];
